+++ src/main/kotlin/com/dprint/formatter/DprintExternalFormatter.kt
@@ -10,6 +10,7 @@ import com.intellij.formatting.service.AsyncFormattingRequest
 import com.intellij.formatting.service.FormattingService
 import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.logger
+import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiFile
 import java.util.concurrent.CancellationException
 import java.util.concurrent.CompletableFuture
@@ -37,8 +38,9 @@ class DprintExternalFormatter : AsyncDocumentFormattingService() {
         // formatter. This is a minor issue and should be resolved if they run it again.
         //
         // We need to take this approach as it appears that blocking the
-        return file.virtualFile != null && file.project.service<EditorServiceManager>()
-            .canFormatCached(file.virtualFile.path) != false
+        val virtualFile = file.virtualFile ?: file.originalFile.virtualFile
+        return virtualFile != null && file.project.service<EditorServiceManager>()
+            .canFormatCached(virtualFile.path) != false
     }
 
     override fun createFormattingTask(formattingRequest: AsyncFormattingRequest): FormattingTask? {
@@ -93,7 +95,7 @@ class DprintExternalFormatter : AsyncDocumentFormattingService() {
                                     path,
                                     resultContent,
                                     range.startOffset,
-                                    range.endOffset,
+                                    getEndOfRange(content, range),
                                     nextHandler
                                 )
                             }
@@ -103,16 +105,21 @@ class DprintExternalFormatter : AsyncDocumentFormattingService() {
                     }
                 }
 
-                val initialRange = if (ranges.size > 0) ranges.first() else null
+                // If the version can't range format we always return null, in this case the underlying process will
+                // extract the start byte position (0) and end byte position (content.encodeToByteArray().size) for the
+                // whole file
+                val initialRange =
+                    if (editorServiceManager.canRangeFormat() && ranges.size > 0) ranges.first() else null
                 val initialHandler: (FormatResult) -> Unit = {
                     baseFormatFuture.complete(it)
                 }
+
                 editorServiceManager.format(
                     formattingId,
                     path,
                     content,
-                    initialRange?.startOffset ?: 0,
-                    initialRange?.endOffset ?: content.length,
+                    initialRange?.startOffset,
+                    getEndOfRange(content, initialRange),
                     initialHandler
                 )
                 // Timeouts are handled at the EditorServiceManager level and an empty result will be
@@ -173,6 +180,18 @@ class DprintExternalFormatter : AsyncDocumentFormattingService() {
             override fun isRunUnderProgress(): Boolean {
                 return true
             }
+        }
+    }
+
+    /**
+     * This function gets around an issue where IntelliJ will sometimes send in a formatting range that
+     * is greater than the actual length of the file. Doing this will cause a no-op in dprint for a format.
+     */
+    private fun getEndOfRange(content: String, range: TextRange?): Int? {
+        return when {
+            range == null -> null
+            range.endOffset > content.length -> content.length
+            else -> range.endOffset
         }
     }