+++ src/main/kotlin/com/dprint/services/editorservice/v5/EditorServiceV5.kt
@@ -25,10 +25,6 @@ class EditorServiceV5(val project: Project) : EditorService {
     private val pendingMessages = PendingMessages()
 
     private fun createStdoutListener(): Thread {
-        if (stdoutListener != null) {
-            stdoutListener?.interrupt()
-        }
-
         return thread {
             StdoutListener(editorProcess, pendingMessages).run()
         }
@@ -39,6 +35,11 @@ class EditorServiceV5(val project: Project) : EditorService {
             Bundle.message("editor.service.initialize", getName()), project, LOGGER
         )
         dropMessages()
+        if (stdoutListener != null) {
+            stdoutListener?.interrupt()
+            stdoutListener = null
+        }
+
         editorProcess.initialize()
         stdoutListener = createStdoutListener()
     }
@@ -95,6 +96,8 @@ class EditorServiceV5(val project: Project) : EditorService {
     }
 
     override fun canRangeFormat(): Boolean {
+        // TODO before we can enable this we need to ensure that the formatting indexes passed into fmt are converted
+        //  from string index to byte index correctly
         return false
     }
 
@@ -109,6 +112,7 @@ class EditorServiceV5(val project: Project) : EditorService {
         LogUtils.info(Bundle.message("formatting.file", filePath), project, LOGGER)
         val message = Message(formatId ?: getNextMessageId(), MessageType.FormatFile)
         message.addString(filePath)
+        // TODO We need to properly handle string index to byte index here
         message.addInt(startIndex ?: 0) // for range formatting add starting index
         message.addInt(endIndex ?: content.encodeToByteArray().size) // add ending index
         message.addInt(0) // Override config