+++ src/main/kotlin/com/dprint/services/editorservice/v5/StdoutListener.kt
@@ -4,6 +4,7 @@ import com.dprint.core.Bundle
 import com.dprint.services.editorservice.EditorProcess
 import com.intellij.openapi.diagnostic.logger
 import org.apache.lucene.util.ThreadInterruptedException
+import java.nio.BufferUnderflowException
 
 private const val SLEEP_TIME = 500L
 
@@ -21,6 +22,10 @@ class StdoutListener(private val editorProcess: EditorProcess, private val pendi
             try {
                 handleStdout()
             } catch (e: ThreadInterruptedException) {
+                LOGGER.info(e)
+                return
+            } catch (e: BufferUnderflowException) {
+                // Happens when the editor service is shut down while this thread is waiting to read output
                 LOGGER.info(e)
                 return
             } catch (e: Exception) {